Fixed a problem with FITS files parsing where signed numbers are not correctly converted to unsigned. Added ability to simulate an occultation with the video modelling and to control the maximum distance.

+++ Tangra 3/Helpers/FITSHelper.cs
@@ -172,7 +172,7 @@ namespace Tangra.Helpers
 
 				for (int x = 0; x < width; x++)
 				{
-					uint val = (uint)dataRow[x];
+					uint val = (ushort)dataRow[x];
 
 					data[x, y] = val;
 					medianCalcList.Add(val);
